more personality x2

+++ src/buttons/stop.js
@@ -1,5 +1,8 @@
 const { EmbedBuilder } = require('discord.js');
-module.exports = async ({ client, inter, queue }) => { 
+module.exports = async ({ client, inter, queue }) => {
+    if (!queue.node.queue.channel.members.has(inter.member.id)) {
+        return inter.editReply({ content: `You are not in a voice channel ${inter.member}... try again ? ❌`, ephemeral: true });
+    }
     if (!queue || !queue.isPlaying()) return inter.editReply({ content: `No music currently playing... try again ? ❌`, ephemeral: true });
 
     queue.delete();
